Due to initialization order, need to forward class-specific values
Tables are a PITA in any GUI toolkit! Due to initialization order of the table cell class hierarchy, the class-specific values for frequency editing are not yet available when the local override of the method to make the associated TextField is called, so we have to forward those values in the local constructor after the superconstructor is called.